♻️ Refactor onboarding form component
